# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Predictions With Sats API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CurrentMarket {
  /** @nullable */
  windowId: number | null;
  /** open, closed, or none */
  status: string;
  /** Current BTC price in USD */
  btcPriceUsd: number;
  /**
   * BTC price when window opened
   * @nullable
   */
  openPrice: number | null;
  /** Seconds until window closes */
  secondsRemaining: number;
  /** Total sats bet on UP */
  totalUpSats: number;
  /** Total sats bet on DOWN */
  totalDownSats: number;
  /** @nullable */
  closesAt: string | null;
}

/**
 * Bet direction
 */
export type CreateBetBodyDirection =
  (typeof CreateBetBodyDirection)[keyof typeof CreateBetBodyDirection];

export const CreateBetBodyDirection = {
  up: "up",
  down: "down",
} as const;

/**
 * Which crypto market to bet on
 */
export type CreateBetBodyAsset =
  (typeof CreateBetBodyAsset)[keyof typeof CreateBetBodyAsset];

export const CreateBetBodyAsset = {
  btc: "btc",
  eth: "eth",
  sol: "sol",
} as const;

export interface CreateBetBody {
  /**
   * Bet amount in USD (converted to sats by server)
   * @minimum 0.5
   */
  amountUsd: number;
  /** Bet direction */
  direction: CreateBetBodyDirection;
  /** Which crypto market to bet on */
  asset: CreateBetBodyAsset;
}

export type BetResponseDirection =
  (typeof BetResponseDirection)[keyof typeof BetResponseDirection];

export const BetResponseDirection = {
  up: "up",
  down: "down",
} as const;

export interface BetResponse {
  id: number;
  paymentHash: string;
  /** BOLT11 Lightning invoice */
  paymentRequest: string;
  amountSats: number;
  direction: BetResponseDirection;
  expiresAt: string;
  windowId: number;
}

export type BetStatusDirection =
  (typeof BetStatusDirection)[keyof typeof BetStatusDirection];

export const BetStatusDirection = {
  up: "up",
  down: "down",
} as const;

export interface BetStatus {
  id: number;
  paymentHash: string;
  direction: BetStatusDirection;
  amountSats: number;
  /** pending, paid, won, lost, expired */
  status: string;
  /** @nullable */
  payoutSats: number | null;
  windowId: number;
  createdAt: string;
  /** @nullable */
  paidAt: string | null;
  /**
   * LNURL-Withdraw token, present when status="won"
   * @nullable
   */
  withdrawToken?: string | null;
  /**
   * unclaimed | claimed
   * @nullable
   */
  withdrawStatus?: string | null;
  /** @nullable */
  withdrawLnurl?: string | null;
  /** @nullable */
  windowOutcome?: string | null;
}

export interface SettledWindow {
  id: number;
  /**
   * up, down, or draw
   * @nullable
   */
  outcome: string | null;
  /** @nullable */
  openPrice: number | null;
  /** @nullable */
  closePrice: number | null;
  /** @nullable */
  priceChangePercent: number | null;
  totalUpSats: number;
  totalDownSats: number;
  openedAt: string;
  /** @nullable */
  settledAt: string | null;
}

export interface PlatformStats {
  totalBets: number;
  totalVolumeSats: number;
  totalWindowsSettled: number;
  /** Fraction of settled windows where UP won (0-1) */
  upWinRate: number;
}

export interface WebhookResponse {
  ok: boolean;
}

export interface ErrorResponse {
  error: string;
}

export type GetMarketHistoryParams = {
  limit?: number;
};

export type AlbyWebhookBody = { [key: string]: unknown };