# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Predictions With Sats API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bet:
    post:
      operationId: createBet
      tags: [bet]
      summary: Place a bet
      description: Creates a bet and returns a Lightning invoice for payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBetBody"
      responses:
        "201":
          description: Bet created with Lightning invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BetResponse"
        "400":
          description: Invalid request or window closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bet/{paymentHash}:
    get:
      operationId: getBetStatus
      tags: [bet]
      summary: Get bet status
      description: Returns the current status of a bet by its payment hash
      parameters:
        - name: paymentHash
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bet status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BetStatus"
        "404":
          description: Bet not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    CreateBetBody:
      type: object
      properties:
        amountUsd:
          type: number
          description: Bet amount in USD (converted to sats by server)
          minimum: 0.5
        direction:
          type: string
          enum: [up, down]
          description: Bet direction
        asset:
          type: string
          enum: [btc, eth, sol]
          description: Which crypto market to bet on
      required:
        - amountUsd
        - direction
        - asset
    BetResponse:
      type: object
      properties:
        id:
          type: integer
        paymentHash:
          type: string
        paymentRequest:
          type: string
          description: BOLT11 Lightning invoice
        amountSats:
          type: integer
        direction:
          type: string
          enum: [up, down]
        expiresAt:
          type: string
          format: date-time
        windowId:
          type: integer
      required:
        - id
        - paymentHash
        - paymentRequest
        - amountSats
        - direction
        - expiresAt
        - windowId
    BetStatus:
      type: object
      properties:
        id:
          type: integer
        paymentHash:
          type: string
        direction:
          type: string
          enum: [up, down]
        amountSats:
          type: integer
        status:
          type: string
          description: "pending, paid, won, lost, expired"
        payoutSats:
          type: ["integer", "null"]
        windowId:
          type: integer
        createdAt:
          type: string
          format: date-time
        paidAt:
          type: ["string", "null"]
          format: date-time
        withdrawToken:
          type: ["string", "null"]
          description: LNURL-Withdraw token, present when status="won"
        withdrawStatus:
          type: ["string", "null"]
          description: unclaimed | claimed
        withdrawLnurl:
          type: ["string", "null"]
        windowOutcome:
          type: ["string", "null"]
      required:
        - id
        - paymentHash
        - direction
        - amountSats
        - status
        - payoutSats
        - windowId
        - createdAt
        - paidAt
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error