# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Predictions With Sats API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns the active betting window with current BTC price and countdown
 * @summary Get current market window
 */
export const GetCurrentMarketResponse = zod.object({
  windowId: zod.number().nullable(),
  status: zod.string().describe("open, closed, or none"),
  btcPriceUsd: zod.number().describe("Current BTC price in USD"),
  openPrice: zod.number().nullable().describe("BTC price when window opened"),
  secondsRemaining: zod.number().describe("Seconds until window closes"),
  totalUpSats: zod.number().describe("Total sats bet on UP"),
  totalDownSats: zod.number().describe("Total sats bet on DOWN"),
  closesAt: zod.coerce.date().nullable(),
});

/**
 * Returns the last settled market windows with outcomes
 * @summary Get recent market history
 */
export const getMarketHistoryQueryLimitDefault = 10;

export const GetMarketHistoryQueryParams = zod.object({
  limit: zod.coerce.number().default(getMarketHistoryQueryLimitDefault),
});

export const GetMarketHistoryResponseItem = zod.object({
  id: zod.number(),
  outcome: zod.string().nullable().describe("up, down, or draw"),
  openPrice: zod.number().nullable(),
  closePrice: zod.number().nullable(),
  priceChangePercent: zod.number().nullable(),
  totalUpSats: zod.number(),
  totalDownSats: zod.number(),
  openedAt: zod.coerce.date(),
  settledAt: zod.coerce.date().nullable(),
});
export const GetMarketHistoryResponse = zod.array(GetMarketHistoryResponseItem);

/**
 * Creates a bet and returns a Lightning invoice for payment
 * @summary Place a bet
 */
export const createBetBodyAmountUsdMin = 0.5;

export const CreateBetBody = zod.object({
  amountUsd: zod
    .number()
    .min(createBetBodyAmountUsdMin)
    .describe("Bet amount in USD (converted to sats by server)"),
  direction: zod.enum(["up", "down"]).describe("Bet direction"),
  asset: zod
    .enum(["btc", "eth", "sol"])
    .describe("Which crypto market to bet on"),
});

/**
 * Returns the current status of a bet by its payment hash
 * @summary Get bet status
 */
export const GetBetStatusParams = zod.object({
  paymentHash: zod.coerce.string(),
});

export const GetBetStatusResponse = zod.object({
  id: zod.number(),
  paymentHash: zod.string(),
  direction: zod.enum(["up", "down"]),
  amountSats: zod.number(),
  status: zod.string().describe("pending, paid, won, lost, expired"),
  payoutSats: zod.number().nullable(),
  windowId: zod.number(),
  createdAt: zod.coerce.date(),
  paidAt: zod.coerce.date().nullable(),
  withdrawToken: zod
    .string()
    .nullish()
    .describe('LNURL-Withdraw token, present when status=\"won\"'),
  withdrawStatus: zod.string().nullish().describe("unclaimed | claimed"),
  withdrawLnurl: zod.string().nullish(),
  windowOutcome: zod.string().nullish(),
});

/**
 * Receives payment confirmation from Alby
 * @summary Alby payment webhook
 */
export const AlbyWebhookBody = zod.record(zod.string(), zod.unknown());

export const AlbyWebhookResponse = zod.object({
  ok: zod.boolean(),
});

/**
 * Returns aggregate stats for the platform
 * @summary Get platform statistics
 */
export const GetPlatformStatsResponse = zod.object({
  totalBets: zod.number(),
  totalVolumeSats: zod.number(),
  totalWindowsSettled: zod.number(),
  upWinRate: zod
    .number()
    .describe("Fraction of settled windows where UP won (0-1)"),
});