# lib/api-zod/src/generated/types/betStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Predictions With Sats API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BetStatusDirection } from "./betStatusDirection";

export interface BetStatus {
  id: number;
  paymentHash: string;
  direction: BetStatusDirection;
  amountSats: number;
  /** pending, paid, won, lost, expired */
  status: string;
  /** @nullable */
  payoutSats: number | null;
  windowId: number;
  createdAt: Date;
  /** @nullable */
  paidAt: Date | null;
  /**
   * LNURL-Withdraw token, present when status="won"
   * @nullable
   */
  withdrawToken?: string | null;
  /**
   * unclaimed | claimed
   * @nullable
   */
  withdrawStatus?: string | null;
  /** @nullable */
  withdrawLnurl?: string | null;
  /** @nullable */
  windowOutcome?: string | null;
}